// src/types/experienceListTypes.ts
export interface Experience {
  id: number;
  userId: number;
  title: string;
  description: string;
  category: string;
  price: number;
  address: string;
  bannerImageUrl: string;
  rating: number;
  reviewCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface ExperienceResponse {
  cursorId: number;
  totalCount: number;
  activities: Experience[];
}